added backend aircons

+++ backend/lights/migrations/20241012160231-create-lights-table.js
@@ -2,10 +2,6 @@
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    // await queryInterface.renameColumn('lights', 'month', 'lights_name');
-    // await queryInterface.renameColumn('lights', 'year', 'lights_unit_usage');
-    // await queryInterface.renameColumn('lights', 'usage', 'lights_count');
-    // await queryInterface.removeColumn('lights', 'price');
     await queryInterface.createTable('lights', {
       id: {
         type: Sequelize.INTEGER,
@@ -21,7 +17,7 @@ module.exports = {
         allowNull: false,
       },
       lights_unit_usage: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL,
         allowNull: false,
       },
       lights_count: {
